combine template engine and transform to reduce confusion
